1.异常查询util优化成连接池 2.dwd层添加批量添加mysql

+++ biz-module-flink-pro/src/main/java/com/yqwl/datamiddle/realtime/bean/DwmSptb02.java
@@ -1045,7 +1045,7 @@ public class DwmSptb02 implements Serializable {
      * 20210422 DTF 大众标准物流时间   number 转  bigint
      */
 
-    private Long NBZWLSJ_DZ;
+    private Double NBZWLSJ_DZ;
 
     /**
      * 20210422 DTF 大众标准到货时间   date 统一时间戳 bigint
@@ -1271,7 +1271,6 @@ public class DwmSptb02 implements Serializable {
      * dwm到货是否正常 1正常0异常
      */
     private Integer SITE_FLAG;
-
 
 
     @Override
